Minor refactor for CleanupNotifier.
